feat(oxidize-c): generation engine + mesh election + k8s + rdma + scrutiny + vision encoder/preprocess/prompt + node + tok/s CLI
New modules (92K lines, 1658 tests):
- generation.c: token generation config/state/result with timing
- election.c: Raft-style leader election (26 tests)
- k8s.c: Kubernetes integration stub (15 tests)
- rdma.c: RDMA zero-copy stub (15 tests)
- scrutiny.c: output validation (banned phrases, PII, length, 20 tests)
- vision_encoder.c: CLIP-style vision encoder stub (19 tests)
- vision_preprocess.c: image resize/center_crop/normalize (20 tests)
- vision_prompt.c: multimodal prompt (LLaVA/Qwen-VL/InternVL/mPLUG-Owl, 20 tests)
- node.c: mesh node identity/connections/traffic (23 tests)

CLI: tok/s timing now printed to stderr after generation
Fixes: numa format-truncation, turboquant fabs, chat_template zero-length,
gradient unused silu, oxk_avx512 unused variable

+++ oxidize-c/src/compute/turboquant.c
@@ -63,7 +63,7 @@ OcError oc_turboquant_calibrate(OcTurboQuant *tq, const float *data, size_t n)
         /* Compute the mean absolute value to detect skew. */
         double sum_abs = 0.0;
         for (size_t i = 0; i < tq->calibration_count; i++) {
-            sum_abs += fabsf((double)tq->calibration_data[i]);
+            sum_abs += fabs((double)tq->calibration_data[i]);
         }
         double mean_abs = sum_abs / (double)tq->calibration_count;
         /* If the data is highly skewed (mean << max), boost scales to
